Remove result loader component
