+++ src/ensembl/src/content/app/custom-download/containers/content/filter-accordion/FiltersAccordion.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { connect } from 'react-redux';
 import { RootState } from 'src/store';
